+++ core/dbt/task/runnable.py
@@ -23,6 +23,7 @@ from dbt.compilation import compile_manifest
 
 from dbt.contracts.graph.compiled import CompileResultNode
 from dbt.contracts.graph.manifest import Manifest
+from dbt.contracts.graph.parsed import ParsedSourceDefinition
 from dbt.contracts.results import ExecutionResult
 from dbt.exceptions import (
     InternalException,
@@ -117,9 +118,17 @@ class GraphRunnableTask(ManifestTask):
                                                     selected_nodes)
 
         # we use this a couple times. order does not matter.
-        self._flattened_nodes = [
-            self.manifest.nodes[uid] for uid in selected_nodes
-        ]
+        self._flattened_nodes = []
+        for uid in selected_nodes:
+            if uid in self.manifest.nodes:
+                self._flattened_nodes.append(self.manifest.nodes[uid])
+            elif uid in self.manifest.sources:
+                self._flattened_nodes.append(self.manifest.sources[uid])
+            else:
+                raise InternalException(
+                    f'Node selection returned {uid}, expected a node or a '
+                    f'source'
+                )
 
         self.num_nodes = len([
             n for n in self._flattened_nodes
@@ -191,7 +200,7 @@ class GraphRunnableTask(ManifestTask):
         """If the caller has passed the magic 'single-threaded' flag, call the
         function directly instead of pool.apply_async. The single-threaded flag
          is intended for gathering more useful performance information about
-        what appens beneath `call_runner`, since python's default profiling
+        what happens beneath `call_runner`, since python's default profiling
         tools ignore child threads.
 
         This does still go through the callback path for result collection.
@@ -266,7 +275,10 @@ class GraphRunnableTask(ManifestTask):
         if self.manifest is None:
             raise InternalException('manifest was None in _handle_result')
 
-        self.manifest.update_node(node)
+        if isinstance(node, ParsedSourceDefinition):
+            self.manifest.update_source(node)
+        else:
+            self.manifest.update_node(node)
 
         if result.error is not None:
             if is_ephemeral: